moved receiver policy to receiver.d.ts/receiver.js
