+++ send_receive/lib/client.d.ts
@@ -8,32 +8,11 @@ import Receiver = require('./receiver');
 import Sender = require('./sender');
 
 declare namespace EventHubClient {
-    enum receiverSettleMode {
-        autoSettle,
-        settleOnDisposition
-    }
-
-    interface ReceiverFlowControlPolicy {
-        /**
-         * Auto or On disposition
-         * One of EventHubClient.receiverSettleMode.autoSettle or EventHubClient.receiverSettleMode.settleOnDisposition
-         */
-        receiverSettleMethod: receiverSettleMode;
-        /**
-         * Re-crediting policy if not auto-settling
-         */
-        creditPolicy?: () => void;
-        /**
-         * Initial number of credits for the receiver if not auto-settling
-         */
-        creditQuantum?: number;
-    }
-
     interface ReceiverOptions {
         startAfterTime?: Date | number;
         startAfterOffset?: string;
         customFilter?: string;
-        flowControlPolicy?: ReceiverFlowControlPolicy;
+        flowControlPolicy?: Receiver.ReceiverFlowControlPolicy;
     }
     type PartitionId = string | number;
 }
